Moved vue files in 'Home' directory so it matches the route structure. Started to restructure register class
