+++ JWT.Website/public/index.html
@@ -12,7 +12,7 @@
             crossorigin="anonymous"
         >
         <!-- Offline use -->
-        <!-- <link rel="stylesheet" href="/bootstrap.css"> -->
+        <link rel="stylesheet" href="/bootstrap.css">
         <link 
             href="https://fonts.googleapis.com/css?family=Montserrat:400,700"
             rel="stylesheet"
