+++ JWT.Website/src/views/Home/RegenerateConfirmationEmail.vue
