+++ JWT.Website/src/views/Home/ResetPassword.vue
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <!-- TODO: Add password reset form and functionality in backend -->
+    </div>
+</template>
