--- JWT.Website/src/views/Home/User/ResetPassword.vue
@@ -1,5 +1,0 @@
-<template>
-    <div>
-        
-    </div>
-</template>
